Mostly completed Application Model implementation. Some testing. Still requires updates of examples and projects, as well as metric work

+++ src/edu/uwo/csd/dcsim/application/loadbalancer/EqualShareLoadBalancer.java
@@ -16,7 +16,7 @@ public class EqualShareLoadBalancer extends LoadBalancer {
 			throw new RuntimeException("Load Balancer called with null Task");
 		}
 		
-		return 1 / task.getInstances().size();
+		return 1 / (float)task.getInstances().size();
 	}
